Upload Ansible collection for Distil (#1)
This is an Ansible collection for deploying and managing Distil, a rating service for OpenStack.

For more information on how to use it, refer to the `README` file.

This is the initial version being released to GitHub, and tagged as `1.0.0`. If this starts getting usage by people as-is, I'd also like to upload it to Ansible Galaxy.